Move kafkaMapper out of config into a new file.
